Center vertically in addition to horizontally
Extending on #12 by introducing the same method to vertically center the
modal. The vertical center used is one based on the golden ratio to make
it look more balanced than an exact 50% vertical center.

I understand that this is a breaking change, but perhaps it can be of
interest still. The one thing in my initial impression that struck me as
a bit odd was that the modal wasn't centered vertically.

For now I will just use the `modalStyles` option to add this centering
myself to my own modals.